fix(ci): move describe-auth-issue into the auth domain lib and restore the login JS budget
components/** is .tsx-only; describe-auth-issue.ts had no JSX and belongs
next to auth-state-mapping.ts under lib/domain/auth. Also inline the two
private helpers into a single mapper (fewer minified bytes, same copy) and
raise the /login JS gzip-9 cap 490000 -> 491000 B for the genuine feature
bytes this PR adds (previously-invisible auth errors now surface on the
login screen).
