+++ apps/packages/product-client/src/components/auth/AuthShell.tsx
@@ -1,7 +1,7 @@
 import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
 
 import { AuthScreenLayout } from "#product/components/auth/AuthScreenLayout";
-import { describeAuthIssue } from "#product/components/auth/describe-auth-issue";
+import { describeAuthIssue } from "#product/lib/domain/auth/describe-auth-issue";
 import { useGitHubSignIn } from "#product/hooks/auth/workflows/use-github-sign-in";
 import { usePasswordSignIn } from "#product/hooks/auth/workflows/use-password-sign-in";
 import { useSsoSignIn } from "#product/hooks/auth/workflows/use-sso-sign-in";
